ai-npc-wormman/server/requirements.txt
fastapi
uvicorn[standard]
websockets
httpx
kokoro-onnx
soundfile
faster-whisper
SpeechRecognition
PyAudio
numpy
imageio-ffmpeg
Pillow
